+++ src/main/java/wac/miles/springBootTest/repository/HouseRepository.java
@@ -3,5 +3,5 @@ package wac.miles.springBootTest.repository;
 import org.springframework.data.repository.CrudRepository;
 import wac.miles.springBootTest.model.House;
 
-public interface HouseRepository  extends CrudRepository<House, Integer> {
+public interface HouseRepository  extends CrudRepository<House, Long> {
 }
